refactor: move solana internals into workspace package
Extract self-contained Solana domain logic from apps/sdp-api into a new
private @sdp/solana workspace package (address helpers, decimal amount
utilities, Token-2022 service + pure utils), mirroring the @sdp/rpc
extraction. Old app import paths remain as thin re-export shims so no
consumer rewrites are needed.
