+++ apps/sdp-api/package.json
@@ -44,6 +44,7 @@
     "@react-email/render": "2.0.8",
     "@sdp/env-config": "workspace:*",
     "@sdp/rpc": "workspace:*",
+    "@sdp/solana": "workspace:*",
     "@sdp/types": "workspace:*",
     "@sentry/cloudflare": "^10.57.0",
     "@sentry/node": "^10.57.0",
